SBERDOMA-616 rename redisGuard to guards
